Changed elevation to double

+++ lib/tap_card.dart
@@ -4,7 +4,7 @@ class MyCard extends Card {
 
   GestureTapCallback onTap;
 
-  MyCard({Key key, Color color, int elevation = 2, Widget child, this.onTap}) : super(key: key, color: color, elevation: elevation, child: child);
+  MyCard({Key key, Color color, double elevation = 2.0, Widget child, this.onTap}) : super(key: key, color: color, elevation: elevation, child: child);
 
   @override
   Widget build(BuildContext context) {
